Fix: dependencies issue resolved
The UI was also wrong showing the onboarding screen instead of the signup/login page. Thw expo router has been resolved now it is showing the signup/login page but still has some issues, the user cannot create an account becauese theure are not fields and when the user tries to log in it automatically fails because there is not call to the database to verify the user infomration. This has to be resolved first before the user can get into the app because there is no back or exit button. This process may well be the end of me.